Finish test move over for simple files
