+++ api/tests/unit/organisations/test_unit_organisations_tasks.py
@@ -1,15 +1,79 @@
 import uuid
 from datetime import timedelta
 
+import pytest
 from django.utils import timezone
 
+from organisations.chargebee.metadata import ChargebeeObjMetadata
 from organisations.models import (
     Organisation,
     OrganisationRole,
     UserOrganisation,
 )
-from organisations.tasks import finish_subscription_cancellation
+from organisations.subscriptions.constants import (
+    FREE_PLAN_ID,
+    MAX_SEATS_IN_FREE_PLAN,
+)
+from organisations.subscriptions.xero.metadata import XeroSubscriptionMetadata
+from organisations.tasks import (
+    ALERT_EMAIL_MESSAGE,
+    ALERT_EMAIL_SUBJECT,
+    finish_subscription_cancellation,
+    send_org_over_limit_alert,
+)
 from users.models import FFAdminUser
+
+
+def test_send_org_over_limit_alert_for_organisation_with_free_subscription(
+    organisation, mocker
+):
+    # Given
+    mocked_ffadmin_user = mocker.patch("organisations.tasks.FFAdminUser")
+
+    # When
+    send_org_over_limit_alert(organisation.id)
+
+    # Then
+    args, kwargs = mocked_ffadmin_user.send_alert_to_admin_users.call_args
+    assert len(args) == 0
+    assert len(kwargs) == 2
+    assert kwargs["message"] == ALERT_EMAIL_MESSAGE % (
+        organisation.name,
+        organisation.num_seats,
+        MAX_SEATS_IN_FREE_PLAN,
+        FREE_PLAN_ID,
+    )
+    assert kwargs["subject"] == ALERT_EMAIL_SUBJECT
+
+
+@pytest.mark.parametrize(
+    "SubscriptionMetadata", [ChargebeeObjMetadata, XeroSubscriptionMetadata]
+)
+def test_send_org_over_limit_alert_for_organisation_with_subscription(
+    organisation, subscription, mocker, SubscriptionMetadata
+):
+    # Given
+    mocked_ffadmin_user = mocker.patch("organisations.tasks.FFAdminUser")
+    max_seats = 10
+    mocker.patch(
+        "organisations.tasks.get_subscription_metadata",
+        return_value=SubscriptionMetadata(seats=max_seats),
+    )
+
+    # When
+    send_org_over_limit_alert(organisation.id)
+
+    # Then
+    args, kwargs = mocked_ffadmin_user.send_alert_to_admin_users.call_args
+    assert len(args) == 0
+    assert len(kwargs) == 2
+    assert kwargs["message"] == ALERT_EMAIL_MESSAGE % (
+        organisation.name,
+        organisation.num_seats,
+        max_seats,
+        subscription.plan,
+    )
+    assert kwargs["subject"] == ALERT_EMAIL_SUBJECT
 
 
 def test_finish_subscription_cancellation(db: None):
